Xp fix and new system

+++ Assets/Scripts/Stat.cs
@@ -18,7 +18,7 @@ public class Stat {
 		Level = 1;
     }
 
-    public void addXP(int value)
+    public void addXP(float value)
     {
 		XP += (int)(value * Multiplier);
 		Level = 1 + XP / XP_TO_LEVEL;
